+++ src/decorator/after.decorator.ts
@@ -1,5 +1,5 @@
-import { BeforeMiddleware } from '../middleware';
-import { BeforeOption } from '../metadata';
+import { AfterMiddleware } from '../middleware';
+import { AfterOption } from '../metadata';
 import { getMetadataArgsStorage } from '../metadata-builder';
 
 /**
@@ -7,7 +7,12 @@ import { getMetadataArgsStorage } from '../metadata-builder';
  *
  * Usage example:
  * ```
- * let middleware: BeforeMiddleware = (context: any, args: { [key: string]: any }, next: (error?: Error, value?: any) => any): any => {
+ * let middleware: AfterMiddleware = (
+ *   context: any,
+ *   args: { [key: string]: any },
+ *   result: Promise<any> | any,
+ *   next: (error?: Error, value?: any) => any
+ * ): any => {
  *   if(context.user.role != 'any role') {
  *     // can use context and resolve/return value as `1000`, for example, regardless of what resolve function actually implements
  *     next(null, 1000);
@@ -21,7 +26,7 @@ import { getMetadataArgsStorage } from '../metadata-builder';
  *
  * class ObjType {
  *   @Field()
- *   @Before(middleware)
+ *   @After(middleware)
  *   myFunction(input: number): number {
  *     return input * 2;
  *   }
@@ -30,15 +35,15 @@ import { getMetadataArgsStorage } from '../metadata-builder';
  * ```
  * @param option Options for an Schema
  */
-export function Before(option: BeforeOption | BeforeMiddleware) {
+export function After(option: AfterOption | AfterMiddleware) {
   return function (target: any, propertyKey: any, index: number) {
-    getMetadataArgsStorage().befores.push({
+    getMetadataArgsStorage().afters.push({
       target: target,
       name: target.name || propertyKey,
-      description: option && (option as BeforeOption).description ? (option as BeforeOption).description : null,
+      description: option && (option as AfterOption).description ? (option as AfterOption).description : null,
       index: index,
       property: propertyKey,
-      middleware: option && (option as BeforeOption).middleware ? (option as BeforeOption).middleware : option as BeforeMiddleware,
+      middleware: option && (option as AfterOption).middleware ? (option as AfterOption).middleware : option as AfterMiddleware,
     });
   } as Function;
 }