+++ src/metadata/args/after.arg.ts
@@ -1,0 +1,8 @@
+import { AfterMiddleware } from '../../middleware';
+import { Argument } from './argument';
+
+export interface AfterArg extends Argument {
+  index: number;
+  property: string;
+  middleware: AfterMiddleware;
+}