+++ src/metadata/options/index.ts
@@ -8,3 +8,4 @@ export * from './argument.option';
 export * from './order-by.option';
 export * from './before.option';
 export * from './interface.option';
+export * from './after.option';