+++ src/metadata/types/index.ts
@@ -11,3 +11,4 @@ export * from './order-by.metadata';
 export * from './root.metadata';
 export * from './field.metadata';
 export * from './interface.metadata';
+export * from './after.metadata';