+++ src/middleware.ts
@@ -1,1 +1,0 @@
-export type Middleware = (context: any, args: { [key: string]: any }, next: (error?: Error, value?: any) => any) => Promise<any> | any;